Refactor: Pull sendNackToQueue up to BaseEhrHandler
Can be reused between tests.
